Pull out playground search styles
