+++ styles/modules/Playground.module.scss
@@ -125,67 +125,9 @@ $GAP: 1rem;
   font-weight: 700;
 }
 
-.searchCard {
-  width: $CARD_WIDTH;
-  height: fit-content;
-  padding: 1rem;
-  margin: 0.5rem auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.cardHeader {
-  text-align: left;
-  @include ellipsis(calc(100vw - 7rem));
-  white-space: pre-line;
-
-  @include tabletAndLarger {
-    max-width: calc(100vw - 24rem);
-  }
-}
-
-.cardText {
-  @extend .cardHeader;
-  font-size: 0.8rem;
-  margin-left: 0.5rem;
-  @include ellipsis(calc(100vw - 7rem));
-
-  @include tabletAndLarger {
-    max-width: calc(100vw - 24rem);
-  }
-}
-
-.actionButton,
-.selectedButton {
-  margin-top: 0.25rem;
-  align-self: flex-end;
-}
-
 .selectionContainer {
   padding: 1rem 0.75rem;
   width: $CARD_WIDTH;
-}
-
-.searchContainer {
-  width: 95%;
-  margin: 0.5rem auto;
-  height: fit-content;
-
-  @include tabletAndLarger {
-    margin: 1rem auto 2rem;
-  }
-}
-
-.viewAllButton {
-  font-size: 0.85rem;
-  min-height: fit-content;
-  padding: 0.25rem;
-  margin-left: auto;
-}
-
-.viewMoreButton {
-  margin: 0.5rem auto;
 }
 
 .codeContent {
@@ -238,26 +180,6 @@ $GAP: 1rem;
     box-shadow: $DT_BOX_SHADOW;
   }
 
-  .searchCard {
-    background-color: $DT_BACKGROUND_MEDIUM;
-    box-shadow: $DT_BOX_SHADOW;
-  }
-
-  .cardHeader {
-    color: $DT_MAIN_TEXT;
-  }
-
-  .cardText {
-    color: $DT_SUB_TEXT;
-  }
-
-  .selectedButton {
-    svg,
-    p {
-      color: $BRIGHT_GREEN;
-    }
-  }
-
   .copyVersionButton {
     @include transition;
     color: $WHITE;
@@ -283,26 +205,6 @@ $GAP: 1rem;
   .resourceDrawerNavigation {
     background-color: $LT_BACKGROUND_LIGHT;
     box-shadow: $LT_BOX_SHADOW;
-  }
-
-  .searchCard {
-    background-color: $LT_BACKGROUND_LIGHT;
-    box-shadow: $LT_BOX_SHADOW;
-  }
-
-  .cardHeader {
-    color: $LT_MAIN_TEXT;
-  }
-
-  .cardText {
-    color: $LT_SUB_TEXT;
-  }
-
-  .selectedButton {
-    svg,
-    p {
-      color: $GREEN;
-    }
   }
 
   .copyVersionButton {
