+++ styles/modules/PlaygroundSearchAndResults.module.scss
@@ -1,0 +1,128 @@
+/**
+ * Copyright 2021 The Rode Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "styles/constants";
+@import "styles/mixins";
+
+$CARD_WIDTH: 97%;
+$GAP: 1rem;
+
+.selectionContainer {
+  padding: 1rem 0.75rem;
+  width: $CARD_WIDTH;
+}
+
+.searchContainer {
+  width: 95%;
+  margin: 0.5rem auto;
+  height: fit-content;
+
+  @include tabletAndLarger {
+    margin: 1rem auto 2rem;
+  }
+}
+
+.viewAllButton {
+  font-size: 0.85rem;
+  min-height: fit-content;
+  padding: 0.25rem;
+  margin-left: auto;
+}
+
+.searchCard {
+  width: $CARD_WIDTH;
+  height: fit-content;
+  padding: 1rem;
+  margin: 0.5rem auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.cardHeader {
+  text-align: left;
+  @include ellipsis(calc(100vw - 7rem));
+  white-space: pre-line;
+
+  @include tabletAndLarger {
+    max-width: calc(100vw - 24rem);
+  }
+}
+
+.cardText {
+  @extend .cardHeader;
+  font-size: 0.8rem;
+  margin-left: 0.5rem;
+  @include ellipsis(calc(100vw - 7rem));
+
+  @include tabletAndLarger {
+    max-width: calc(100vw - 24rem);
+  }
+}
+
+.actionButton,
+.selectedButton {
+  margin-top: 0.25rem;
+  align-self: flex-end;
+}
+
+.viewMoreButton {
+  margin: 0.5rem auto;
+}
+
+.darkTheme {
+  .searchCard {
+    background-color: $DT_BACKGROUND_MEDIUM;
+    box-shadow: $DT_BOX_SHADOW;
+  }
+
+  .cardHeader {
+    color: $DT_MAIN_TEXT;
+  }
+
+  .cardText {
+    color: $DT_SUB_TEXT;
+  }
+
+  .selectedButton {
+    svg,
+    p {
+      color: $BRIGHT_GREEN;
+    }
+  }
+}
+
+.lightTheme {
+  .searchCard {
+    background-color: $LT_BACKGROUND_LIGHT;
+    box-shadow: $LT_BOX_SHADOW;
+  }
+
+  .cardHeader {
+    color: $LT_MAIN_TEXT;
+  }
+
+  .cardText {
+    color: $LT_SUB_TEXT;
+  }
+
+  .selectedButton {
+    svg,
+    p {
+      color: $GREEN;
+    }
+  }
+}
